Add standalone in true. Update routing
