+++ src/app/app.component.ts
@@ -3,12 +3,11 @@ import { Component } from '@angular/core';
 import { LayoutComponent } from "@presentation/shared/components/layout/layout.component";
 
 @Component({
+  standalone: true,
   selector: 'app-root',
   imports: [LayoutComponent],
-  template: `<layout-page />`,
+  template: `<app-layout />`,
 })
 export class AppComponent {
-
   title = 'Business Card IngAamira';
-
 }
