--- src/domain/index.ts
@@ -1,1 +1,0 @@
-// Barrel export file
