+++ src/presentation/features/portfolio/pages/not-found/not-found.component.ts
@@ -1,0 +1,18 @@
+import { Component } from '@angular/core';
+import { Title } from '@angular/platform-browser';
+import { RouterLink } from '@angular/router';
+
+@Component({
+  standalone: true,
+  selector: 'app-not-found',
+  imports: [RouterLink],
+  templateUrl: './not-found.component.html',
+  styleUrls: ['./not-found.component.css'],
+})
+export class NotFoundComponent {
+
+  constructor (private titleService: Title) {
+    this.titleService.setTitle('404')
+  }
+
+}
