+++ src/presentation/features/portfolio/pages/project-modal/project-modal.component.ts
@@ -1,0 +1,20 @@
+import { CommonModule } from '@angular/common';
+import { Component} from '@angular/core';
+
+import { BsModalRef } from 'ngx-bootstrap/modal';
+import { CarouselModule } from 'ngx-bootstrap/carousel';
+
+import { Project } from '@presentation/shared/interfaces/project';
+
+@Component({
+  standalone: true,
+  selector: 'app-project-modal',
+  imports: [CommonModule, CarouselModule],
+  templateUrl: './project-modal.component.html'
+})
+export class ProjectModalComponent {
+  project = {} as Project;
+
+  constructor(public bsModalRef: BsModalRef) { }
+
+}
