--- src/presentation/features/profile/pages/not-found/not-found.component.ts
@@ -1,17 +1,0 @@
-import { Component } from '@angular/core';
-import { Title } from '@angular/platform-browser';
-import { RouterLink } from '@angular/router';
-
-@Component({
-  selector: 'app-not-found',
-  imports: [RouterLink],
-  templateUrl: './not-found.component.html',
-  styleUrls: ['./not-found.component.css'],
-})
-export default class NotFoundComponent {
-
-  constructor (private titleService: Title) {
-    this.titleService.setTitle('404')
-  }
-
-}
