--- src/presentation/features/profile/pages/project-modal/project-modal.component.ts
@@ -1,19 +1,0 @@
-import { CommonModule } from '@angular/common';
-import { Component} from '@angular/core';
-
-import { BsModalRef } from 'ngx-bootstrap/modal';
-import { CarouselModule } from 'ngx-bootstrap/carousel';
-
-import { Project } from '@presentation/shared/interfaces/project';
-
-@Component({
-  selector: 'app-project-modal',
-  imports: [CommonModule, CarouselModule],
-  templateUrl: './project-modal.component.html'
-})
-export class ProjectModalComponent {
-  project = {} as Project;
-
-  constructor(public bsModalRef: BsModalRef) { }
-
-}
