+++ src/presentation/shared/components/layout/layout.component.ts
@@ -2,11 +2,12 @@ import { Component } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { CommonModule } from '@angular/common';
 
-import { HeaderComponent } from './../../../shared/components/header/header.component';
-import { NavbarComponent } from './../../../shared/components/navbar/navbar.component';
-import { FooterComponent } from './../../../shared/components/footer/footer.component';
+import { HeaderComponent } from '@presentation/shared/components/header/header.component';
+import { NavbarComponent } from '@presentation/shared/components/navbar/navbar.component';
+import { FooterComponent } from '@presentation/shared/components/footer/footer.component';
 
 @Component({
+  standalone: true,
   selector: 'app-layout',
   imports: [
     CommonModule,
