+++ tsconfig.json
@@ -5,9 +5,10 @@
   "compilerOptions": {
     "baseUrl": "./",
     "paths": {
-        "@domain/*": ["src/domain/*"],
-        "@infrastructure/*": ["src/infrastructure/*"],
-        "@presentation/*": ["src/presentation/*"]
+      "@domain/*": ["src/domain/*"],
+      "@infrastructure/*": ["src/infrastructure/*"],
+      "@presentation/*": ["src/presentation/*"],
+      "@assets/*": ["src/assets/*"],
     },
     "outDir": "./dist/out-tsc",
     "strict": true,
